make ts-loader ignore declaration files
