+++ packages/nextjs-lambda-origin-request/webpack.config.js
@@ -7,7 +7,11 @@ module.exports = {
       {
         test: /\.tsx?$/,
         use: "ts-loader",
-        exclude: /node_modules/
+        exclude: /node_modules|\.d\.ts$/
+      },
+      {
+        test: /\.d\.ts$/,
+        loader: "ignore-loader"
       }
     ]
   },
